requirements.txt
django
uvicorn
django-ckeditor
asgiref
python-dateutil
six
sqlparse
tzdata